HW5 end

+++ HW5/Adapter/ISquare.php
@@ -1,5 +1,5 @@
 <?php
 interface ISquare
 {
-function squareArea(int $sideSquare);
+function squareArea(float $sideSquare);
 }
